pyproject.toml
[project]
name = "musync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "spotipy>=2.24.0",
    "ytmusicapi>=1.8.2",
]
